tester/pyproject.toml
[project]
name = "voice-tester-agent"
version = "0.1.0"
description = "Test caller agent for voice agent testing"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.4.0",
    "livekit-plugins-cartesia",
    "livekit-plugins-deepgram",
    "livekit-plugins-openai",
    "livekit-plugins-silero",
    "python-dotenv",
    "pyyaml",
    "aiohttp",
]
